+++ firmware/src/main/GSwifi.h
@@ -183,7 +183,7 @@ public:
      * HTTP request
      * @param timeout_second closes connection and dispatches callback with status_code 498 after this number of times caller calls onTimer()
      */
-    int8_t request(GSMETHOD method, const char *path, const char *body, uint8_t length, GSResponseHandler handler, uint8_t timeout_second, uint8_t is_binary);
+    int8_t request(GSMETHOD method, const char *path, const char *body, uint16_t length, GSResponseHandler handler, uint8_t timeout_second, uint8_t is_binary);
     int8_t get (const char *path, GSResponseHandler handler, uint8_t timeout_second);
     int8_t postBinary (const char *path, const char *body, uint16_t length, GSResponseHandler handler, uint8_t timeout_second);
     int8_t post (const char *path, const char *body, uint16_t length, GSResponseHandler handler, uint8_t timeout_second);
